scripts: fix path detecting for colorscheme.sh

+++ scripts/colorscheme.sh
@@ -23,21 +23,24 @@ if [ "$THEME"x == ""x ]; then
 fi
 
 _switch_color_scheme() {
-    [ -f ~/.config/nvim/envs.vim ] && \
+    if [ -f ~/.config/nvim/envs.vim ]; then
         sed -E -i.bak \
-        "s/let g:vim_color_scheme = '.*'/let g:vim_color_scheme = '$2'/" \
-        ~/.config/nvim/envs.vim && \
-        rm ~/.config/nvim/envs.vim.bak
-    [ -f ~/.tmux.conf ] && \
+            "s/let g:vim_color_scheme = '.*'/let g:vim_color_scheme = '$2'/" \
+            ~/.config/nvim/envs.vim && \
+            rm ~/.config/nvim/envs.vim.bak
+    fi
+    if [ -f ~/.tmux.conf ]; then
         sed -E -i.bak \
-        "s/^source.*/source ~\/\.tmux\/tmuxline\/$1\.tmux\.conf/" \
-        ~/.tmux.conf && \
-        rm ~/.tmux.conf.bak
-    [ -f ~/.config/zathura/zathurarc ] && \
+            "s/^source.*/source ~\/\.tmux\/tmuxline\/$1\.tmux\.conf/" \
+            ~/.tmux.conf && \
+            rm ~/.tmux.conf.bak
+    fi
+    if [ -f ~/.config/zathura/zathurarc ]; then
         sed -E -i.bak \
-        "s/^include themes.*/include themes\/$1/" \
-        ~/.config/zathura/zathurarc && \
-        rm ~/.config/zathura/zathurarc.bak
+            "s/^include themes.*/include themes\/$1/" \
+            ~/.config/zathura/zathurarc && \
+            rm ~/.config/zathura/zathurarc.bak
+    fi
     rm ~/.zsh-theme
     cp -f "${3}.zsh" ~/.zsh-theme
 }
